requirements.txt
Flask==2.3.3
requests==2.31.0
gspread==5.11.3
oauth2client==4.1.3
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pandas==2.1.4
numpy==1.24.3
Pillow==10.1.0
python-dotenv==1.0.0
matplotlib==3.8.2
pytz==2023.3.post1
python-dateutil==2.8.2
gunicorn==21.2.0
geopy==2.4.0  